Compiler warning
git-svn-id: http://svn.ta3d.org/yuni/trunk@1622 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/tools/yuni-docmake/job.cpp
@@ -208,7 +208,7 @@ namespace // anonymous
 				if (!text.empty())
 				{
 					StringAdapter adapter;
-					adapter.adapt(text.c_str(), text.size());
+					adapter.adapt(text.c_str(), (unsigned int) text.size());
 					seo(adapter);
 				}
 				break;
